Update to the Circular Buffer
The Circular Buffer should now be functional. Removed the ICollection<T>
interface as it did not apply to the definition of a circular buffer as
a FIFO data structure. Summary comments for some of the methods have
also been added. The class may be able to use a little more work.